Remove github CI on pull_request
The pull_request event on github is always detached HEAD and the
GITHUB_REF is also unset. The commit SHA is even not accessible
through CLI git commands, so versioning from that is not possible in a
meaningful and portable way.

Builds should be done on branches if mvn deploy is ought to make sense.