More enforcement
Enforce Java versions prior to 9 without messing with the compiler
flags that maven.compiler.release then brings. This should be cleaned
up if Java 8 is not supported anymore.

Require release dependencies and upper bounds deps. There is no way to
require versions on plugins as that misses jar and resources plugin
which is probably a bug in the enforcer.

Add local emacs variables to the pom for less indenting flippy floppy.